# Baatchit_v1.0/src/main/java/com/baatchit/common/StringResponse.java
package com.baatchit.common;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class StringResponse {

    private String response;
}